Fix the Travis CI matrix
By specifying all of the Python interpretters at the root of the
configuration, I blew out the number of build cases. An alternative, so
that we're only testing py27 against the Python 2.7 interpretter seems
to be to define the matrix at a higher level, rather than within the
environment variables.